[C2RKMpiDec]: fix int32_t overflow while testing CtsMediaMiscTestCases
err log:
tombstoned: received crash request for pid 747
tombstoned: crash socket received short read of length 0 (expected 12)

Type: Fix
Redmine ID: N/A
Associated modifications: N/A
Test: CtsMediaMiscTestCases:
android.media.misc.cts.ResourceManagerTest
android.media.misc.cts.ResourceManagerMultiTest

+++ component/osal/include/C2RKMemTrace.h
@@ -55,8 +55,8 @@ private:
     virtual ~C2RKMemTrace() {};
 
     int32_t             mDisableCheck;
-    int32_t             mCurDecLoad;
-    int32_t             mCurEncLoad;
+    int64_t             mCurDecLoad;
+    int64_t             mCurEncLoad;
     uint32_t            mMaxInstanceNum;
     Mutex               mLock;
     Vector<C2NodeInfo>  mDecNodes;
